+++ changelog.md
@@ -5,6 +5,7 @@
 
 ### Added
 * `[utils]` Dump logs to a file on RT1051 after reaching 80% of log buffer utilization.
+* `[call functions]` Window for MMI/USSD internal messages.
 
 ### Changed
 * `[bluetooth]` Underlying communication with the Bluetooth module over DMA (direct access)